added nav bar to other pages, integrated formatting, fixed child tree page rendering
