+++ client/package.json
@@ -14,7 +14,7 @@
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
         "react-hook-form": "^7.54.2",
-        "react-icons": "^5.4.0",
+        "react-icons": "^5.5.0",
         "react-router-dom": "^7.1.3",
         "react-scripts": "5.0.1",
         "reactjs-popup": "^2.0.6",
