+++ client/src/App.js
@@ -7,6 +7,7 @@ import Home from './pages/Home/Home';
 import Tree from './pages/Tree/Tree';
 import ShareTree from './pages/Tree/ShareTree/ShareTree';
 import ViewSharedTrees from './pages/Tree/ViewSharedTrees/ViewSharedTrees';
+import { Outlet } from 'react-router-dom';
 //ToDo: Import a file for app settings, help, and chat page
 //ToDo: Add routes for app settings, help, and chat page
