+++ client/src/pages/Account/styles.js
@@ -1,6 +1,6 @@
 export const DefaultStyle = {
     display: 'flex',
     flexDirection: 'row',
-    padding: '10px',
+    textAlign: 'center',
     fontFamily: 'Alata'
 }
