+++ client/src/pages/Home/Home.js
@@ -6,11 +6,15 @@ import { ReactComponent as CalendarIcon } from '../../assets/calendar.svg';
 import { ReactComponent as PlusIcon } from '../../assets/plus-sign.svg';
 import CreateEventPopup from '../../components/CreateEvent/CreateEvent';
 import CreateMemoryPopup from '../../components/CreateMemory/CreateMemory';
+import NavBar from '../../components/NavBar/NavBar';
 
 // TODO: need to implement auth check for this page (and others), redirect to login if not authenticated
 function Home() {
+    document.body.style.overflow = 'hidden';
+    document.body.style.width = '100%'; 
     return (
         <div style={styles.DefaultStyle}>
+            <NavBar />
             <Link to="/tree" >View Tree Page</Link>
             <CreateEventPopup trigger={<CalendarIcon style={styles.CalendarButton} />} />
             <CreateMemoryPopup trigger={<PlusIcon style={styles.PlusButton} />} />
