+++ client/src/pages/Home/styles.js
@@ -1,7 +1,7 @@
 export const DefaultStyle = {
     display: 'flex',
     flexDirection: 'row',
-    padding: '10px',
+    textAlign: 'center',
     fontFamily: 'Alata'
 }
