--- client/src/pages/ShareTree/styles.js
@@ -1,92 +1,0 @@
-export const DefaultStyle = {
-    display: 'flex',
-    flexDirection: 'column',
-    padding: '10px',
-    fontFamily: 'Alata',
-    alignItems: 'center',
-  }
-
-  export const ContainerStyle = {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'flex-start',
-    width: '40%',
-    maxWidth: '600px',
-    height: 'auto',
-    borderRadius: '30px',
-    backgroundColor: '#faf9f6',
-    padding: '20px',
-    boxShadow: '0px 0px 10px rgba(0, 0, 0, 0.1)',
-    marginTop: '5vh'
-}
-  export const FormStyle = {
-    fontFamily: 'Aboreto',
-    width: '100%',
-    display: 'flex',
-    alignItems: 'center',
-    flexDirection: 'column',
-  }
-
-  export const ListStyle = {
-    listStyleType: 'none',
-    fontFamily: 'Alata',
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    width: '80%',
-    padding: '0'
-  };
-
-  export const ItemStyle = {
-    margin: '10px 0px',
-    width: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-};
-
-export const FieldStyle = {
-    borderRadius: '5px',
-    border: '1px solid #000000',
-    width: '100%',
-    height: '25px',
-};
-
-export const TextAreaStyle = {
-    borderRadius: '5px',
-    border: '1px solid #000000',
-    height: '70px',
-    width: '100%',
-    resize: 'none'
-};
-
-export const ButtonStyle = {
-    fontFamily: 'Alata',
-    backgroundColor: '#3a5a40',
-    color: 'white',
-    borderRadius: '20px',
-    border: 'none',
-    padding: '10px 30px',
-    margin: '10px',
-    cursor: 'pointer',
-}
-
-export const ListingStyle = {
-    padding: '10px', 
-    border: '1px solid gray', 
-    width: '90%'
-}
-
-export const AddOptionsStyle = {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    width: '90%',
-    padding: '10px',
-    fontFamily: 'Alata',
-    // border: '1px solid gray',
-    // borderRadius: '5px',
-    marginTop: '10px',
-    height: '200px',
-    overflow: 'auto'
-}
