--- client/src/pages/Tree/ShareTree/ShareTree.css
@@ -1,6 +1,0 @@
-.sharetree-page {
-    display: flex;
-    text-align: center;
-    color: black;
-    padding-left: 300px;
-}
