--- client/src/pages/Tree/ViewSharedTrees/ViewSharedTrees.css
@@ -1,6 +1,0 @@
-.viewshared-page {
-    display: flex;
-    text-align: center;
-    color: black;
-    padding-left: 300px;
-}
