+++ client/src/pages/Tree/ViewSharedTrees/styles.js
@@ -1,0 +1,43 @@
+export const DefaultStyle = {
+    display: 'flex',
+    flexDirection: 'column',
+    padding: '10px',
+    fontFamily: 'Alata',
+    alignItems: 'center',
+    width: '100%',
+  }
+
+  export const ContainerStyle = {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+    width: '60%',
+    maxWidth: '600px',
+    height: 'auto',
+    borderRadius: '30px',
+    backgroundColor: '#faf9f6',
+    padding: '20px',
+    boxShadow: '0px 0px 10px rgba(0, 0, 0, 0.1)',
+    marginTop: '5vh'
+}
+
+export const ItemStyle = {
+    margin: '10px 0px',
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    border: '1px solid rgb(87, 87, 87)',
+    padding: '10px 20px',
+    alignItems: 'center',
+};
+
+export const ListStyle = {
+    listStyleType: 'none',
+    fontFamily: 'Alata',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    width: '80%',
+    padding: '0'
+  };
